Graphics.Gloss

Description

Gloss hides the pain of drawing simple vector graphics behind a nice data type and a few display functions.

Getting something on the screen is as easy as:

    import Graphics.Gloss
    main = display (InWindow "Nice Window" (200, 200) (10, 10)) white (Circle 80)

Once the window is open you can use the following:

Quit - esc-key.

Move Viewport - left-click drag, arrow keys.

Rotate Viewport - right-click drag, control-left-click drag, or home/end-keys.

Zoom Viewport - mouse wheel, or page up/down-keys.

Animations can be constructed similarly using the animate.

If you want to run a simulation based around finite time steps then try simulate.

If you want to manage your own key/mouse events then use play.

Gloss uses OpenGL under the hood, but you don't have to worry about any of that.

Gloss programs should be compiled with -threaded, otherwise the GHC runtime will limit the frame-rate to around 20Hz.

Release Notes:

For 1.7.0:
   * Tweaked circle level-of-detail reduction code.
   * Increased frame rate cap to 100hz.
   Thanks to Doug Burke
   * Primitives for drawing arcs and sectors.
   Thanks to Thomas DuBuisson
   * IO versions of animate, simplate and play.

For 1.6.0:
   Thanks to Anthony Cowley
   * Full screen display mode.

For 1.5.0:
   * O(1) Conversion of ForeignPtrs to bitmaps.
   * An extra flag on the Bitmap constructor allows bitmaps to be cached
     in texture memory between frames.

For 1.4.0:
   Thanks to Christiaan Baaij: 
   * Refactoring of Gloss internals to support multiple window manager backends.
   * Support for using GLFW as the window library instead of GLUT.
     GLUT is still the default, but to use GLFW install gloss with:
        cabal install gloss --flags="GLFW -GLUT"

For more information, check out http://gloss.ouroborus.net.

Documentation

module Graphics.Gloss.Data.Picture

module Graphics.Gloss.Data.Color

data Display

Describes how Gloss should display its output.

Constructors

InWindow String (Int, Int) (Int, Int)
Display in a window with the given name, size and position.

FullScreen (Int, Int)
Display full screen with a drawing area of the given size.

display

Arguments

:: Display
Display mode.

-> Color
Background color.

-> Picture
The picture to draw.

-> IO ()

Open a new window and display the given picture.

Use the following commands once the window is open:

Quit - esc-key.

Move Viewport - left-click drag, arrow keys.

Rotate Viewport - right-click drag, control-left-click drag, or home/end-keys.

Zoom Viewport - mouse wheel, or page up/down-keys.

animate

Arguments

:: Display
Display mode.

-> Color
Background color.

-> (Float -> Picture)
Function to produce the next frame of animation. It is passed the time in seconds since the program started.

-> IO ()

Open a new window and display the given animation.

Once the window is open you can use the same commands as with display.

simulate

Arguments

:: forall model .

=> Display
Display mode.

-> Color
Background color.

-> Int
Number of simulation steps to take for each second of real time.

-> model
The initial model.

-> (model -> Picture)
A function to convert the model to a picture.

-> (ViewPort -> Float -> model -> model)
A function to step the model one iteration. It is passed the current viewport and the amount of time for this simulation step (in seconds).

-> IO ()

Run a finite-time-step simulation in a window. You decide how the model is represented, how to convert the model to a picture, and how to advance the model for each unit of time. This function does the rest.

Once the window is open you can use the same commands as with display.

play

Arguments

:: forall world .

=> Display
Display mode.

-> Color
Background color.

-> Int
Number of simulation steps to take for each second of real time.

-> world
The initial world.

-> (world -> IO Picture)
A function to convert the world a picture.

-> (Event -> world -> world)
A function to handle input events.

-> (Float -> world -> world)
A function to step the world one iteration. It is passed the period of time (in seconds) needing to be advanced.

-> IO ()

Play a game in a window.