Graphics.Gloss.Data.Color

Description

Predefined and custom colors.

Color data type

data Color

An abstract color value. We keep the type abstract so we can be sure that the components are in the required range. To make a custom color use makeColor.

Constructors

RGBA !Float !Float !Float !Float
Holds the color components. All components lie in the range [0..1.

makeColor

Arguments

:: Float
Red component.

-> Float
Green component.

-> Float
Blue component.

-> Float
Alpha component.

-> Color

Make a custom color. All components are clamped to the range [0..1].

makeColor' :: Float -> Float -> Float -> Float -> Color

Make a custom color. You promise that all components are clamped to the range [0..1]

makeColor8

Arguments

:: Int
Red component.

-> Int
Green component.

-> Int
Blue component.

-> Int
Alpha component.

-> Color

Make a custom color. All components are clamped to the range [0..255].

rawColor

Arguments

:: Float
Red component.

-> Float
Green component.

-> Float
Blue component.

-> Float
Alpha component.

-> Color

Make a custom color. Components should be in the range [0..1] but this is not checked.

rgbaOfColor :: Color -> (Float, Float, Float, Float)

Take the RGBA components of a color.

Color functions

mixColors

Arguments

:: Float
Ratio of first color.

-> Float
Ratio of second color.

-> Color
First color.

-> Color
Second color.

-> Color
Resulting color.

Mix two colors with the given ratios.

addColors :: Color -> Color -> Color

Add RGB components of a color component-wise, then normalise them to the highest resulting one. The alpha components are averaged.

dim :: Color -> Color

Make a dimmer version of a color, scaling towards black.

bright :: Color -> Color

Make a brighter version of a color, scaling towards white.

light :: Color -> Color

Lighten a color, adding white.

dark :: Color -> Color

Darken a color, adding black.

Pre-defined colors

greyN

Arguments

:: Float
Range is 0 = black, to 1 = white.

-> Color

A greyness of a given magnitude.

black, white :: Color

Primary

red, blue, green :: Color

Secondary

yellow, magenta, cyan :: Color

Tertiary

rose, orange, chartreuse, aquamarine, azure, violet :: Color