Graphics.Gloss.Data.Display

Documentation

data Display

Describes how Gloss should display its output.

Constructors

InWindow String (Int, Int) (Int, Int)
Display in a window with the given name, size and position.

FullScreen (Int, Int)
Display full screen with a drawing area of the given size.