Graphics.Gloss.Data.Point

Documentation

type Point = (Float, Float)

A point on the x-y plane. Points can also be treated as Vectors, so Graphics.Gloss.Data.Vector may also be useful.

pointInBox :: Point -> Point -> Point -> Bool

Test whether a point lies within a rectangular box that is oriented on the x-y plane. The points P1-P2 are opposing points of the box, but need not be in a particular order.

    P2 +-------+
       |       |
       | + P0  |
       |       |
       +-------+ P1