Graphics.Gloss.Interface.Pure.Game

Description

This game mode lets you manage your own input. Pressing ESC will still abort the program, but you don't get automatic pan and zoom controls like with displayInWindow.

Documentation

module Graphics.Gloss.Data.Display

module Graphics.Gloss.Data.Picture

module Graphics.Gloss.Data.Color

play

Arguments

:: forall world .

=> Display
Display mode.

-> Color
Background color.

-> Int
Number of simulation steps to take for each second of real time.

-> world
The initial world.

-> (world -> IO Picture)
A function to convert the world a picture.

-> (Event -> world -> world)
A function to handle input events.

-> (Float -> world -> world)
A function to step the world one iteration. It is passed the period of time (in seconds) needing to be advanced.

-> IO ()

Play a game in a window.

data Event

Possible input events.

Constructors

EventKey Key KeyState Modifiers (Float, Float)

EventMotion (Float, Float)

data Key

Constructors

Char Char

SpecialKey SpecialKey

MouseButton MouseButton

data SpecialKey

Constructors

KeyUnknown

KeySpace

KeyEsc

KeyF1

KeyF2

KeyF3

KeyF4

KeyF5

KeyF6

KeyF7

KeyF8

KeyF9

KeyF10

KeyF11

KeyF12

KeyF13

KeyF14

KeyF15

KeyF16

KeyF17

KeyF18

KeyF19

KeyF20

KeyF21

KeyF22

KeyF23

KeyF24

KeyF25

KeyUp

KeyDown

KeyLeft

KeyRight

KeyTab

KeyEnter

KeyBackspace

KeyInsert

KeyNumLock

KeyBegin

KeyDelete

KeyPageUp

KeyPageDown

KeyHome

KeyEnd

KeyPad0

KeyPad1

KeyPad2

KeyPad3

KeyPad4

KeyPad5

KeyPad6

KeyPad7

KeyPad8

KeyPad9

KeyPadDivide

KeyPadMultiply

KeyPadSubtract

KeyPadAdd

KeyPadDecimal

KeyPadEqual

KeyPadEnter

data MouseButton

Constructors

LeftButton

MiddleButton

RightButton

WheelUp

WheelDown

AdditionalButton Int

data KeyState

Constructors

Down

Up

data Modifiers

Constructors

Modifiers

Fields

shift :: KeyState

ctrl :: KeyState

alt :: KeyState