Graphics.Gloss.Internals.Interface.Animate

Documentation

animateWithBackendIO

Arguments

:: Backend a

=> a
Initial State of the backend

-> Bool
Whether to allow the image to be panned around.

-> Display
Display mode.

-> Color
Background color.

-> (Float -> IO Picture)
Function to produce the next frame of animation. It is passed the time in seconds since the program started.

-> IO ()