Graphics.Gloss.Internals.Interface.Backend.Types

Documentation

class Backend a where

The functions every backend window managed backend needs to support.

The Backend module interfaces with the window manager, and handles opening and closing the window, and managing key events etc.

It doesn't know anything about drawing lines or setting colors. When we get a display callback, Gloss will perform OpenGL actions, and the backend needs to have OpenGL in a state where it's able to accept them.

Methods

initBackendState :: a

Initialize the state used by the backend. If you don't use any state, make a Unit-like type; see the GLUT backend for an example.

initializeBackend :: IORef a -> Bool -> IO ()

Perform any initialization that needs to happen before opening a window The Boolean flag indicates if any debug information should be printed to the terminal

exitBackend :: IORef a -> IO ()

Perform any deinitialization and close the backend.

openWindow :: IORef a -> Display -> IO ()

Open a window with the given display mode.

dumpBackendState :: IORef a -> IO ()

Dump information about the backend to the terminal.

installDisplayCallback :: IORef a -> [Callback] -> IO ()

Install the display callbacks.

installWindowCloseCallback :: IORef a -> IO ()

Install the window close callback.

installReshapeCallback :: IORef a -> [Callback] -> IO ()

Install the reshape callbacks.

installKeyMouseCallback :: IORef a -> [Callback] -> IO ()

Install the keymouse press callbacks.

installMotionCallback :: IORef a -> [Callback] -> IO ()

Install the mouse motion callbacks.

installIdleCallback :: IORef a -> [Callback] -> IO ()

Install the idle callbacks.

runMainLoop :: IORef a -> IO ()

The mainloop of the backend.

postRedisplay :: IORef a -> IO ()

A function that signals that screen has to be updated.

getWindowDimensions :: IORef a -> IO (Int, Int)

Function that returns (width,height) of the window in pixels.

elapsedTime :: IORef a -> IO Double

Function that reports the time elapsed since the application started. (in seconds)

sleep :: IORef a -> Double -> IO ()

Function that puts the current thread to sleep for n seconds.

type DisplayCallback = forall a. Backend a => IORef a -> IO ()

Display callback has no arguments.

type KeyboardMouseCallback = forall a. Backend a => IORef a -> Key -> KeyState -> Modifiers -> (Int, Int) -> IO ()

Arguments: KeyType, Key Up / Down, Ctrl / Alt / Shift pressed, latest mouse location.

type MotionCallback = forall a. Backend a => IORef a -> (Int, Int) -> IO ()

Arguments: (PosX,PosY) in pixels.

type IdleCallback = forall a. Backend a => IORef a -> IO ()

No arguments.

type ReshapeCallback = forall a. Backend a => IORef a -> (Int, Int) -> IO ()

Arguments: (Width,Height) in pixels.

data Callback

Constructors

Display DisplayCallback

KeyMouse KeyboardMouseCallback

Idle IdleCallback

Motion MotionCallback

Reshape ReshapeCallback

data Key

Constructors

Char Char

SpecialKey SpecialKey

MouseButton MouseButton

data MouseButton

Constructors

LeftButton

MiddleButton

RightButton

WheelUp

WheelDown

AdditionalButton Int

data KeyState

Constructors

Down

Up

data SpecialKey

Constructors

KeyUnknown

KeySpace

KeyEsc

KeyF1

KeyF2

KeyF3

KeyF4

KeyF5

KeyF6

KeyF7

KeyF8

KeyF9

KeyF10

KeyF11

KeyF12

KeyF13

KeyF14

KeyF15

KeyF16

KeyF17

KeyF18

KeyF19

KeyF20

KeyF21

KeyF22

KeyF23

KeyF24

KeyF25

KeyUp

KeyDown

KeyLeft

KeyRight

KeyTab

KeyEnter

KeyBackspace

KeyInsert

KeyNumLock

KeyBegin

KeyDelete

KeyPageUp

KeyPageDown

KeyHome

KeyEnd

KeyPad0

KeyPad1

KeyPad2

KeyPad3

KeyPad4

KeyPad5

KeyPad6

KeyPad7

KeyPad8

KeyPad9

KeyPadDivide

KeyPadMultiply

KeyPadSubtract

KeyPadAdd

KeyPadDecimal

KeyPadEqual

KeyPadEnter

data Modifiers

Constructors

Modifiers

Fields

shift :: KeyState

ctrl :: KeyState

alt :: KeyState

module Graphics.Gloss.Data.Display