Graphics.Gloss.Internals.Interface.Game

Documentation

playWithBackendIO

Arguments

:: forall world a . Backend a

=> a
Initial state of the backend

-> Display
Display mode.

-> Color
Background color.

-> Int
Number of simulation steps to take for each second of real time.

-> world
The initial world.

-> (world -> IO Picture)
A function to convert the world to a picture.

-> (Event -> world -> IO world)
A function to handle input events.

-> (Float -> world -> IO world)
A function to step the world one iteration. It is passed the period of time (in seconds) needing to be advanced.

-> IO ()

data Event

Possible input events.

Constructors

EventKey Key KeyState Modifiers (Float, Float)

EventMotion (Float, Float)