Graphics.Gloss.Internals.Interface.Simulate

Documentation

simulateWithBackendIO

Arguments

:: forall model a . Backend a

=> a
Initial state of the backend

-> Display
Display mode.

-> Color
Background color.

-> Int
Number of simulation steps to take for each second of real time.

-> model
The initial model.

-> (model -> IO Picture)
A function to convert the model to a picture.

-> (ViewPort -> Float -> model -> IO model)
A function to step the model one iteration. It is passed the current viewport and the amount of time for this simulation step (in seconds).

-> IO ()