Graphics.Gloss.Internals.Interface.ViewPort

Description

The ViewPort represents the global transformation applied to the displayed picture. When the user pans, zooms, or rotates the display then this changes the ViewPort.

Documentation

data ViewPort

Constructors

ViewPort

Fields

viewPortTranslate :: !(Float, Float)

Global translation.

viewPortRotate :: !Float

Global rotation (in degrees).

viewPortScale :: !Float

Global scaling (of both x and y coordinates).

viewPortInit :: ViewPort

The initial state of the viewport.