Interpreter

Documentation

foo :: [Char] -> IO ()

foo accepts commands till no familiar pattern match at which it gives outputimage

data Commands

Handle Operations on the data

data Point

Point as a new data type

data Picture

The Picture Data Type Circle - Center-x, Center-y, Radius (x,y are cartesian co-ordinates) Line - Two tuples of floats (The line between them) Polyline - Line linking a list of points - Can be used to build Square, Rectangle, Triangle or any general Polygon

rotateAndShiftPoint :: (Float, Float) -> Float -> Float -> Float -> (Float, Float)

Handling Rotations and shift of axes - Function #1

shiftAndRotatePoint :: (Float, Float) -> Float -> Float -> Float -> (Float, Float)

Handling Rotations and shift of axes - Function #2

foo1 :: Float -> Float -> Float -> Float -> Float -> (Float, Float)

Supplementary function #1 for generatePoints method - Line case

foo2 :: Float -> Float -> Float -> Float -> Float -> Float -> Bool

Supplementary function #1 for generatePoints method - Ellipse case

generatePoints :: Picture -> [(Float, Float)]

Supplementary function #1 for outputBMP method Code to generate the array of points in a picture given the picture as input

fillColor :: Commands -> IO ()

fillColor fills the figure starting at the point given to it until it encounters a black pixel barrier Uses the flood fill algorithm to fill up blank space in the figure from the starting point DFS approach used

filter_f_1 :: (Float, Float) -> [(Float, Float)] -> [Word8]

Supplementary function #2 for fillColor method

filter_f_2 :: (Float, Float) -> [(Float, Float)] -> [Word8]

Supplementary function #3 for fillColor method

outputBMP :: [Picture] -> [Word8] -> IO ()

Function to write the final bmp image of all the pictures fed by the user The drawing area and hence the size of the bmp image is assumed to be 600*600 pixels Fixed sized image 568 pixels *383 pixels

zeroArray :: Int -> [Word8]

Give an empty array of given length Used to create a blank image

getBinaryArray :: [Picture] -> [Word8]

Supplementary function #2 for outputBMP method Take a list of Pictures and output the vector of raster image data