Main

Documentation

main :: IO ()

The main function. [OPTIONAL] Can take an input file from the command line to start with.

makePicture :: State -> Picture

Convert our state to a picture.

makePicture' :: State -> IO Picture

A wrapper for calling makePicture function.

picRead :: [Picture] -> [Char] -> [Picture]

This function is used to display pictures mentioned ithe input file. Converts strings read from files to actual pictures.

showLn :: Show a => a -> String

Show with a new line

handleEvent :: Event -> State -> State

Handle mouse click and motion events.

data State

The game state

stepWorld :: Float -> State -> State

sdegrees :: Float -> Float -> Float -> Float -> Float -> Float -> Float

Computes the angle (in degrees) between two lines , connected by a given point. Used for square

scaleNormalizer :: Float -> Float -> Float

root mean square of the two floats.

edegrees :: Float -> Float -> Float -> Float -> Float -> Float -> Float

Computes the angle (in degrees) between two lines l1 and l2, connected by (x_cen, y_cen)

marker :: Maybe (Float, Float) -> Picture

Used in Surround function

dist :: (Float, Float) -> (Float, Float) -> Float

Euclidean distance between two points.

between :: (Float, Float) -> (Float, Float) -> (Float, Float) -> Bool

If a point lies between two other points.

closerPoint :: (Float, Float) -> Maybe (Float, Float) -> Maybe (Float, Float) -> Maybe (Float, Float)

Return the point closer to the first argument.

closestPoint :: (Float, Float) -> [Maybe (Float, Float)] -> Maybe (Float, Float)

Finds the closest point in a list of points.

rotateAndShiftPoint :: (Float, Float) -> Float -> Float -> Float -> (Float, Float)

Rotates a given point by a given angle and then translates it by a given distance.

lineInter :: (Float, Float) -> (Float, Float) -> (Float, Float) -> (Float, Float) -> Maybe (Float, Float)

The function that calls intersections to compute line intersections with any shape.

intersections :: (Float, Float) -> (Float, Float) -> Picture -> [Maybe (Float, Float)]

Given a line, returns a set of points intersecting with any shape.

surround :: (Float, Float) -> [Picture] -> [Maybe (Float, Float)]

Used for user ease during color fill. It marks the boundary of to fileed by marker dots

triangleFill :: (Float, Float) -> [Maybe (Float, Float)] -> [Picture]

Fills colors using triangles projecting out from the point of colour filling tap. Tha is why this fills only the region visible from the point of origin/fill command. THIS IS NOT FLOOD FILLING.

colorMaker :: (Float, Float) -> Color

This function assigns colour to fill when user clicks on a colour in the colour pallette.